backup/requirements.txt
pynput
customtkinter



Pillow
mss

numpy

pyautogui
pyperclip
